refactor layout and main, and some nav
